Move project content indexer out into own code
This allows it to be run in isolation without having to also rebuild the main projects index, which allows it to be rebuilt on prod without affecting core functionality.
