+++ lib/indexers/index.js
@@ -5,6 +5,7 @@ const { NotFoundError } = require('@asl/service/errors');
 const indexers = {
   establishments: require('./establishments'),
   projects: require('./projects'),
+  'projects-content': require('./projects-content'),
   profiles: require('./profiles'),
   places: require('./places')
 };
@@ -23,7 +24,12 @@ module.exports = (settings) => {
   app.put('/:index/:id', (req, res, next) => {
     return Promise.resolve(settings.esClient)
       .then(client => {
-        return indexers[req.params.index](db, client, { id: req.params.id });
+        return indexers[req.params.index](db, client, { id: req.params.id })
+          .then(() => {
+            if (req.params.index === 'projects') {
+              return indexers['projects-content'](db, client, { id: req.params.id })
+            }
+          });
       })
       .then(() => {
         res.json({ message: `Re-indexed ${req.params.index}:${req.params.id}` });
