+++ lib/indexers/projects-content/extract-content.js
