+++ lib/indexers/projects-content/index.js
@@ -1,15 +1,12 @@
 const { pick, get } = require('lodash');
-const extractSpecies = require('./extract-species');
+const extractContent = require('./extract-content');
 
-const indexName = 'projects';
+const indexName = 'projects-content';
 const columnsToIndex = [
   'id',
   'title',
   'status',
   'licenceNumber',
-  'expiryDate',
-  'raDate',
-  'isLegacyStub',
   'schemaVersion'
 ];
 
@@ -25,22 +22,21 @@ const indexProject = (esClient, project, ProjectVersion) => {
     .then(version => {
       version = version || {};
       const data = version.data || {};
-      const species = extractSpecies(data, project);
-      return Promise.resolve()
-        .then(() => {
-          return esClient.index({
-            index: indexName,
-            id: project.id,
-            body: {
-              ...pick(project, columnsToIndex),
-              licenceNumber: project.licenceNumber ? project.licenceNumber.toUpperCase() : null,
-              licenceHolder: pick(project.licenceHolder, 'id', 'firstName', 'lastName'),
-              establishment: pick(project.establishment, 'id', 'name'),
-              keywords: data.keywords,
-              species
-            }
-          });
-        });
+      const content = extractContent(data, project);
+      const protocols = data.protocols.map(p => pick(p, 'id', 'title'));
+      return esClient.index({
+        index: indexName,
+        id: project.id,
+        body: {
+          ...pick(project, columnsToIndex),
+          licenceNumber: project.licenceNumber ? project.licenceNumber.toUpperCase() : null,
+          licenceHolder: pick(project.licenceHolder, 'id', 'firstName', 'lastName'),
+          establishment: pick(project.establishment, 'id', 'name'),
+          versionId: version.id,
+          protocols,
+          content
+        }
+      });
     });
 };
 
@@ -99,15 +95,6 @@ const reset = esClient => {
               },
               status: {
                 type: 'keyword',
-                fields: {
-                  value: {
-                    type: 'keyword'
-                  }
-                }
-              },
-              species: {
-                type: 'keyword',
-                normalizer: 'lowercase',
                 fields: {
                   value: {
                     type: 'keyword'
