+++ lib/search/global/projects-content.js
@@ -9,7 +9,7 @@ module.exports = client => async (term = '', query = {}) => {
   };
 
   params.body.query = {
-    bool: {},
+    bool: {}
   };
   params.body.highlight = {
     fields: {
@@ -24,18 +24,10 @@ module.exports = client => async (term = '', query = {}) => {
     }
   };
 
-  if (query.filters || query.species) {
+  if (query.filters && query.filters.status && query.filters.status[0]) {
     const filter = { term: {} };
-
-    if (query.filters && query.filters.status && query.filters.status[0]) {
-      filter.term.status = query.filters.status[0];
-    }
-    if (query.species) {
-      filter.term.species = query.species;
-    }
-    if (Object.keys(filter.term).length) {
-      params.body.query.bool.filter = filter;
-    }
+    filter.term.status = query.filters.status[0];
+    params.body.query.bool.filter = filter;
   }
 
   if (!term) {
